requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
diffusers==0.27.2
transformers==4.40.1
accelerate==0.30.1
safetensors==0.4.3
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
python-multipart==0.0.9
huggingface-hub==0.23.0
